// src/scraper/enums/blog-type.enum.ts
export enum BlogType {
  GOOGLE = 'google',
  AMAZON = 'amazon',
  MICROSOFT = 'microsoft',
  FACEBOOK = 'facebook',
  INSTAGRAM = 'instagram',
  LINKEDIN = 'linkedin',
  NETFLIX = 'netflix',
  TWITTER = 'twitter',
  AIRBNB = 'airbnb',
  UBER = 'uber',
  SLACK = 'slack',
  DROPBOX = 'dropbox',
  TWITCH = 'twitch',
  PINTEREST = 'pinterest',
  EVERNOTE = 'evernote',
}